Created  dir + Split parts of code from pipeline into errorListener
