+++ src/core/pipeline/pipeline.ts
@@ -4,183 +4,30 @@
  */
 
 import { performance } from 'perf_hooks'
-import { isError } from 'util'
+import { CharStreams, CommonTokenStream } from 'antlr4'
+import { isDebug } from '../../config/env'
+import YiniLexer from '../../grammar/generated/YiniLexer'
+import YiniParser, { YiniContext } from '../../grammar/generated/YiniParser'
 import {
-    CharStreams,
-    CommonTokenStream,
-    ErrorListener,
-    RecognitionException,
-    Recognizer,
-    Token,
-} from 'antlr4'
-import {
-    isDebug,
-    isDev,
-    isProdEnv,
-    localAppEnv,
-    localNodeEnv,
-} from '../config/env'
-import YiniLexer from '../grammar/generated/YiniLexer'
-import YiniParser, { YiniContext } from '../grammar/generated/YiniParser'
-import {
-    FailLevelKey,
     ParsedObject,
     ParseOptions,
     ResultMetadata,
     YiniParseResult,
-} from '../types'
-import { removeUndefinedDeep, sortObjectKeys } from '../utils/object'
-import { debugPrint, printObject } from '../utils/print'
-import { toLowerKebabCase, toLowerSnakeCase } from '../utils/string'
-import astBuilder from './astBuilder'
-import { ErrorDataHandler } from './errorDataHandler'
-import { IParseCoreOptions, IRuntimeInfo, IYiniAST } from './internalTypes'
-import { astToObject } from './objectBuilder'
+} from '../../types'
+import { removeUndefinedDeep } from '../../utils/object'
+import { debugPrint, printObject } from '../../utils/print'
+import astBuilder from '../astBuilder'
+import { ErrorDataHandler } from '../errorDataHandler'
+import { IParseCoreOptions, IRuntimeInfo, IYiniAST } from '../internalTypes'
+import { astToObject } from '../objectBuilder'
 import {
     buildResultMetadata,
     IBuildResultMetadataParams,
-} from './resultMetadataBuilder'
-
-/**
- * @param line Line number as 1-based.
- * @param col Column number as 0-based.
- */
-const createGeneralCtx = (
-    line: number,
-    endColumn: number,
-    startColumn: number | undefined = undefined,
-): YiniContext => {
-    const startToken = new Token()
-    const stopToken = new Token()
-    const ctx = new YiniContext()
-    ctx.start = startToken
-    ctx.stop = stopToken
-
-    ctx.start.line = line // Note: Line num is 1-based.
-    if (startColumn && startColumn >= 0) {
-        ctx.start.column = startColumn // Note: Column num is 0-based.
-    }
-
-    ctx.stop.column = endColumn // Note: Column num is 0-based.
-
-    return ctx
-}
-
-const parsePossibleStartCol = (
-    errorHandler: ErrorDataHandler,
-    recognizer: any,
-): number | undefined => {
-    let possibleStartCol: number | undefined = undefined
-    try {
-        possibleStartCol = recognizer?._ctx.start?.column
-            ? recognizer?._ctx.start?.column + 1
-            : undefined
-    } catch (err: unknown) {
-        let msgHint = ''
-
-        if (isError(err)) {
-            msgHint = 'Error: ' + err.message
-        } else {
-            msgHint = 'Thrown value:' + JSON.stringify(err)
-        }
-
-        if (isProdEnv()) {
-            return 0
-        }
-
-        errorHandler.pushOrBail(
-            null,
-            'Internal-Error',
-            'Catched error of possibleStartCol in parser grammar listener.',
-            msgHint,
-        )
-    }
-
-    return possibleStartCol
-}
-
-class MyParserErrorListener implements ErrorListener<any> {
-    public errors: string[] = []
-    private errorHandler: ErrorDataHandler
-
-    constructor(errorHandler: ErrorDataHandler) {
-        this.errorHandler = errorHandler
-    }
-
-    syntaxError(
-        recognizer: any,
-        offendingSymbol: any,
-        line: number,
-        charPositionInLine: number,
-        msg: string,
-        e: RecognitionException | undefined,
-    ): void {
-        debugPrint('ANTLR parser cached an error')
-
-        const col = charPositionInLine + 1
-        const possibleStartCol: number | undefined = parsePossibleStartCol(
-            this.errorHandler,
-            recognizer,
-        )
-
-        const msgWhat = `Syntax error`
-
-        // Try to map message:
-        // From: "mismatched input '/END' expecting <EOF>"
-        // To:   "Found '/END', but expected the end of the document."
-        // const msgWhy = `${capitalizeFirst(msg)}`
-        const msgWhy = `Details: ${msg}`
-
-        const ctx = createGeneralCtx(line, charPositionInLine, possibleStartCol) // So we can show line/col in error message.
-        // Note, after pushing processing may continue or exit, depending on the error and/or the bail threshold.
-        this.errorHandler.pushOrBail(ctx, 'Syntax-Error', msgWhat, msgWhy)
-    }
-
-    // The following are required for the interface, but can be left empty.
-    reportAmbiguity(...args: any[]): void {}
-    reportAttemptingFullContext(...args: any[]): void {}
-    reportContextSensitivity(...args: any[]): void {}
-}
-
-class MyLexerErrorListener implements ErrorListener<any> {
-    public errors: string[] = []
-    private errorHandler: ErrorDataHandler
-
-    constructor(errorHandler: ErrorDataHandler) {
-        this.errorHandler = errorHandler
-    }
-
-    syntaxError(
-        recognizer: any,
-        offendingSymbol: any,
-        line: number,
-        charPositionInLine: number,
-        msg: string,
-        e: RecognitionException | undefined,
-    ) {
-        // Handle the error as you want:
-        debugPrint('ANTLR lexer cached an error')
-        const col = charPositionInLine + 1
-        const possibleStartCol: number | undefined = parsePossibleStartCol(
-            this.errorHandler,
-            recognizer,
-        )
-
-        // const msgWhat = `Syntax error at line ${line}, column ${col}:`
-        const msgWhat = `Syntax error`
-
-        // Try to map message:
-        // From: "mismatched input '/END' expecting <EOF>"
-        // To:   "Found '/END', but expected the end of the document."
-        // const msgWhy = `${capitalizeFirst(msg)}`
-        const msgWhy = `Details: ${msg}`
-        // const msgHint = ``
-
-        const ctx = createGeneralCtx(line, charPositionInLine, possibleStartCol) // So we can show line/col in error message.
-        // Note, after pushing processing may continue or exit, depending on the error and/or the bail threshold.
-        this.errorHandler.pushOrBail(ctx, 'Syntax-Error', msgWhat, msgWhy)
-    }
-}
+} from '../resultMetadataBuilder'
+import {
+    createLexerErrorListener,
+    createParserErrorListener,
+} from './errorListeners'
 
 /**
  * @internal Single source of truth.
@@ -200,7 +47,7 @@ export const runPipeline = (
     _meta_userOpts: ParseOptions,
 ): ParsedObject | YiniParseResult => {
     debugPrint()
-    debugPrint('-> Entered parseMain(..) in parseEntry')
+    debugPrint('-> Entered runPipeline(..) in pipeline.ts')
     debugPrint('    isStrict initialMode = ' + coreOptions.rules.initialMode)
     debugPrint('         bailSensitivity = ' + coreOptions.bailSensitivity)
     debugPrint('           isIncludeMeta = ' + coreOptions.isIncludeMeta)
@@ -287,7 +134,8 @@ export const runPipeline = (
 
     // Remove the default ConsoleErrorListener
     lexer.removeErrorListeners() // Removes the default lexer console error output.
-    const lexerErrorListener = new MyLexerErrorListener(errorHandler)
+    // const lexerErrorListener = new MyLexerErrorListener(errorHandler)
+    const lexerErrorListener = createLexerErrorListener(errorHandler)
     lexer.addErrorListener(lexerErrorListener)
 
     const tokenStream = new CommonTokenStream(lexer)
@@ -313,25 +161,26 @@ export const runPipeline = (
     // const errorListener = new MyParserErrorListener(errorHandler)
 
     parser.removeErrorListeners() // Removes the default parser console error output.
-    const parserErrorListener = new MyParserErrorListener(errorHandler)
+    // const parserErrorListener = new MyParserErrorListener(errorHandler)
+    const parserErrorListener = createParserErrorListener(errorHandler)
     parser.addErrorListener(parserErrorListener)
 
     const parseTree: YiniContext = parser.yini() // The function yini() is the start rule.
-    if (
-        parserErrorListener.errors.length > 0 ||
-        lexerErrorListener.errors.length > 0
-    ) {
-        debugPrint('*** ERROR detected ***')
+    // if (
+    //     parserErrorListener.errors.length > 0 ||
+    //     lexerErrorListener.errors.length > 0
+    // ) {
+    //     debugPrint('*** ERROR detected ***')
 
-        if (isDebug()) {
-            // Handle or display syntax errors
-            console.error(
-                'Syntax errors detected:',
-                parserErrorListener.errors,
-                lexerErrorListener.errors,
-            )
-        }
-    }
+    //     if (isDebug()) {
+    //         // Handle or display syntax errors
+    //         console.error(
+    //             'Syntax errors detected:',
+    //             parserErrorListener.errors,
+    //             lexerErrorListener.errors,
+    //         )
+    //     }
+    // }
 
     debugPrint(
         '=== Ended phase 2 =============================================',
